requirements.txt
# YOLOv8 teeth detection project dependencies

# Core YOLOv8 framework 
ultralytics>=8.2.0

# Data processing and visualization
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning metrics
scikit-learn>=1.0.0

# YAML support (already included with ultralytics but explicitly listed)
PyYAML>=6.0

# Note: OpenCV, Pillow, NumPy, tqdm, requests 
# are automatically installed as ultralytics dependencies
